backend/requirements.txt
fastapi>=0.115.0
numpy>=1.26.0
Pillow==11.0.0
uvicorn[standard]>=0.30.6
python-multipart==0.0.9
openai==1.55.3
python-dotenv==1.0.1
passlib==1.7.4
PyJWT>=2.10.1
mcp[cli]>=1.26.0
aiofiles==24.1.0
